+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- Add context area hover and click interactions (gorarakelyan)
 - Add ability to remove bottom outliers (gorarakelyan)
 - Add ability to remove outliers (roubkar)